fix: repair Stage 4 Windows and Ruff gates
Repair the External Evidence Control Tower link syntax, make the new Stage 4 files Ruff-compliant, keep both project packages first-party for import sorting, and apply the requested uniform medium-grey analytical panel contract. Update regression coverage, version metadata, audit records and portable manifests without changing governed scenario outcomes or the read-only LLM authority boundary.

+++ src/lpr_cpe_demo/digital_twin/install_assurance.py
@@ -226,7 +226,7 @@ def _start_offset_hours(scenario: str) -> float:
 
 
 def _observation_offsets(age_hours: float) -> list[int]:
-    limit = max(0, int(math.floor(age_hours * 60)))
+    limit = max(0, math.floor(age_hours * 60))
     offsets = {0}
     offsets.update(range(5, min(limit, 120) + 1, 5))
     if limit > 120:
